% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/song-print.R
\name{print.song_model}
\alias{print.song_model}
\title{Print a SONG Model}
\usage{
\method{print}{song_model}(x, ...)
}
\arguments{
\item{x}{A \code{"song_model"} object.}

\item{...}{Ignored.}
}
\value{
Invisible \code{x}.
}
\description{
Prints a concise summary of a \code{"song_model"} object.
}
\examples{
model <- song(as.matrix(iris[, 1:4]), epochs = 5L, seed = 42)
print(model)

}
\references{
Senanayake, D. A., Wang, W., Naik, S. H., & Halgamuge, S. (2021).
Self-Organizing Nebulous Growths for Robust and Incremental Data
Visualization. \emph{IEEE Transactions on Neural Networks and Learning
Systems}, 32(10), 4588--4602. \doi{10.1109/TNNLS.2020.3023941}
}
