% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{songR-package}
\alias{songR}
\alias{songR-package}
\title{songR: Self-Organizing Nebulous Growths for Dimensionality Reduction}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Native R/C++ implementation of the Self-Organizing Nebulous Growths (SONG) algorithm for nonlinear dimensionality reduction and data visualization. SONG is a parametric method that supports incremental addition of new data to existing visualizations without reinitialization, handles both homogeneous and heterogeneous data increments, and is robust to noise and highly mixed clusters. Based on the algorithm described in Senanayake et al. (2021) \doi{10.1109/TNNLS.2020.3023941}. Includes an interactive Shiny application for comparing SONG, t-SNE, and UMAP visualizations side-by-side.
}
\references{
Senanayake, D. A., Wang, W., Naik, S. H., & Halgamuge, S. (2021).
Self-Organizing Nebulous Growths for Robust and Incremental Data
Visualization. \emph{IEEE Transactions on Neural Networks and Learning
Systems}, 32(10), 4588--4602. \doi{10.1109/TNNLS.2020.3023941}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-heller.github.io/songR/}
  \item \url{https://github.com/r-heller/songR}
  \item Report bugs at \url{https://github.com/r-heller/songR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

}
\keyword{internal}
